Add Silex.app_class config parameter.
Some Silex installations call `$app->run()` in their `app_path` file,
which will result in the contents of those pages polluting the test
outputs.

This option allows the user to instantiate their AppKernel directly
(which should also handle the route registration processes) so that the
code is easily re-usable.

It may be necessary to add the following to your `composer.json` file
for autoloading of your AppKernel class to be loaded properly:

```yaml
{
    "autoload": {
        "files": [
            "app/AppKernel.php"
        ]
    },
    ...
}
```